requirements.txt
dash
dash-bootstrap-components
plotly
pymupdf
transformers
numpy
pillow
torch
ollama
